# src/docs/swagger.yaml
basePath: /
definitions:
...
  main.APICarRef:
    properties:
      car_id:
        description: CarID TeslaMate cars.id
        type: integer
      car_name:
        description: CarName 车辆昵称，可为空
        type: string
    type: object
  main.APICarsListCar:
    properties:
      car_details:
        allOf:
        - $ref: '#/definitions/main.APICarsListCarDetails'
        description: CarDetails 车型与效率等
      car_exterior:
        allOf:
        - $ref: '#/definitions/main.APICarsListCarExterior'
        description: CarExterior 外观
      car_id:
        description: CarID 车辆 ID
        type: integer
      car_settings:
        allOf:
        - $ref: '#/definitions/main.APICarsListCarSettings'
        description: CarSettings 车辆设置
      name:
        description: Name 昵称
        type: string
      teslamate_details:
        allOf:
        - $ref: '#/definitions/main.APICarsListTMDetails'
        description: TeslaMateDetails 写入/更新时间
      teslamate_stats:
        allOf:
        - $ref: '#/definitions/main.APICarsListTMStats'
        description: TeslaMateStats 充电/驾驶/更新次数统计
    type: object
  main.APICarsListCarDetails:
    properties:
      efficiency:
        allOf:
        - $ref: '#/definitions/main.NullFloat64'
        description: Efficiency 效率系数
      eid:
        description: EID Tesla 车辆 entity id
        type: integer
      model:
        description: Model 车型字符串
        type: string
      trim_badging:
        description: TrimBadging 尾标
        type: string
      vid:
        description: VID 车辆 id
        type: integer
      vin:
        description: Vin 车架号
        type: string
    type: object
  main.APICarsListCarExterior:
    properties:
      exterior_color:
        description: ExteriorColor 颜色
        type: string
      spoiler_type:
        description: SpoilerType 尾翼
        type: string
      wheel_type:
        description: WheelType 轮毂
        type: string
    type: object
  main.APICarsListCarSettings:
    properties:
      free_supercharging:
        description: FreeSupercharging 免费超充
        type: boolean
      req_not_unlocked:
        description: ReqNotUnlocked 要求未解锁
        type: boolean
      suspend_after_idle_min:
        description: SuspendAfterIdleMin 空闲后挂起
        type: integer
      suspend_min:
        description: SuspendMin 挂起分钟
        type: integer
      use_streaming_api:
        description: UseStreamingAPI 流式 API
        type: boolean
    type: object
  main.APICarsListTMDetails:
    properties:
      inserted_at:
        description: InsertedAt 创建时间
        type: string
      updated_at:
        description: UpdatedAt 更新时间
        type: string
    type: object
  main.APICarsListTMStats:
    properties:
      total_charges:
        description: TotalCharges 充电次数
        type: integer
      total_drives:
        description: TotalDrives 驾驶次数
        type: integer
      total_updates:
        description: TotalUpdates 更新次数
        type: integer
    type: object
...
  main.APIChargesBatteryLevels:
    properties:
      end_battery_level:
        type: integer
      start_battery_level:
        type: integer
    type: object
  main.APIChargesRangeSpan:
    properties:
      end_range:
        type: number
      start_range:
        type: number
    type: object
  main.APIChargesRow:
    properties:
      address:
        type: string
      battery_details:
        $ref: '#/definitions/main.APIChargesBatteryLevels'
      charge_energy_added:
        type: number
      charge_energy_used:
        type: number
      charge_id:
        type: integer
      cost:
        type: number
      duration_min:
        type: integer
      duration_str:
        type: string
      end_date:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      odometer:
        type: number
      outside_temp_avg:
        type: number
      range_ideal:
        $ref: '#/definitions/main.APIChargesRangeSpan'
      range_rated:
        $ref: '#/definitions/main.APIChargesRangeSpan'
      start_date:
        type: string
    type: object
...
  main.APIUnitsLengthTemp:
    properties:
      unit_of_length:
        description: UnitsLength km 或 mi
        type: string
      unit_of_temperature:
        description: UnitsTemperature C 或 F
        type: string
    type: object
...
  main.NullFloat64:
    properties:
      float64:
        type: number
      valid:
        description: Valid is true if Float64 is not NULL
        type: boolean
    type: object
...
  main.RespCarsList:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/main.RespCarsListData'
        description: Data 车辆列表或单车
    type: object
  main.RespCarsListData:
    properties:
      cars:
        description: Cars 车辆数组
        items:
          $ref: '#/definitions/main.APICarsListCar'
        type: array
    type: object
...
  main.RespChargesList:
    properties:
      data:
        $ref: '#/definitions/main.RespChargesListData'
    type: object
  main.RespChargesListData:
    properties:
      car:
        $ref: '#/definitions/main.APICarRef'
      charges:
        items:
          $ref: '#/definitions/main.APIChargesRow'
        type: array
      units:
        $ref: '#/definitions/main.APIUnitsLengthTemp'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 与 TeslaMate 采集数据对齐的 JSON API；看板对照见仓库 `执行计划.md`。
  license:
    name: MIT
  title: TeslaMate API
  version: "1.0"
paths:
...
  /api/v1/cars/{CarID}:
    get:
      parameters:
      - description: 车辆 ID；省略则列出全部
        example: 1
        in: path
        name: CarID
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.RespCarsList'
      summary: 车辆列表或详情
      tags:
      - cars
...
  /api/v1/cars/{CarID}/charges:
    get:
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      - description: 开始时间 RFC3339
        in: query
        name: startDate
        type: string
      - description: 结束时间 RFC3339
        in: query
        name: endDate
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 100
        description: 每页条数
        in: query
        name: show
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.RespChargesList'
      summary: 充电记录列表
      tags:
      - charges
...
schemes:
- http
swagger: "2.0"